Move UITests logger class to UITestsFoundation
